operator/multi-pool: Move files under operator/pkg
Multi Pool IPAM mode requires both an operator component as well as an
agent one. Currently, both components, despite being completely
separate, are kept under pkg/ipam/allocator/multipool.  Following the
structure of other packages, move all the operator related logic under
operator/pkg/ipam. This allows to recognize at a glance if a certain
portion of the code is part of the agent or part of the operator.
